+++ Nestjs/src/resource/resource.service.ts
@@ -1,6 +1,7 @@
 import { ICrudServices } from '@/core/crud'
 import { PaginationResponse } from '@/lib/pagination/pagination.interface'
 import { QueryBase, QueryHelper } from '@/lib/query-helper'
+import { PermissionService } from '@/permission/permission.service'
 import { ConflictException, Injectable, Logger, NotFoundException } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { In, Repository } from 'typeorm'
@@ -12,6 +13,7 @@ import { Resource, resourceFields } from './entities/resource.entity'
 export class ResourceService implements ICrudServices {
   private readonly logger = new Logger(ResourceService.name)
   constructor(
+    private readonly permissionService: PermissionService,
     @InjectRepository(Resource) private readonly resourceRepository: Repository<Resource>
   ) {}
 
@@ -20,14 +22,14 @@ export class ResourceService implements ICrudServices {
     return this.resourceRepository.existsBy({ name })
   }
 
-  async create(data: CreateResourceDto): Promise<any> {
+  async create(data: CreateResourceDto) {
     const isExist = await this.checkExistName(data.name)
     if (isExist) throw new ConflictException('Tên resource đã tồn tại')
     const resource = await this.resourceRepository.create(data)
     return this.resourceRepository.save(resource)
   }
 
-  async createMany(data: CreateResourceDto[]): Promise<any> {
+  async createMany(data: CreateResourceDto[]) {
     return this.resourceRepository.save(data)
   }
 
@@ -40,7 +42,7 @@ export class ResourceService implements ICrudServices {
     this.logger.log('Resources Initialized')
   }
 
-  async update(id: number, data: UpdateReourceDto): Promise<any> {
+  async update(id: number, data: UpdateReourceDto) {
     const isExist = await this.checkExistName(data.name)
     if (isExist) throw new ConflictException('Tên resource đã tồn tại')
     const resource = await this.findOne(id)
@@ -48,12 +50,30 @@ export class ResourceService implements ICrudServices {
     return this.resourceRepository.save(resource)
   }
 
-  async delete(id: number): Promise<any> {
-    await this.findOne(id)
+  async delete(id: number) {
+    const resource = await this.findOne(id)
+    const permissionPaginate = await this.permissionService.findMany({
+      filter: { resourceId: { eq: `${id}` } }
+    })
+    const countPermission = permissionPaginate.meta.totalItem
+    if (countPermission > 0)
+      throw new ConflictException(
+        `Không thể xóa, ${resource.name} resource đang được ${countPermission} permission sử dụng`
+      )
     await this.resourceRepository.delete({ id })
   }
 
-  async deleteMany(ids: number[]): Promise<any> {
+  async deleteMany(ids: number[]) {
+    const resourcesCount = await this.resourceRepository.countBy({ id: In(ids) })
+    if (resourcesCount !== ids.length) throw new NotFoundException('Một số resource không tồn tại')
+    const permissionPaginate = await this.permissionService.findMany({
+      filter: { resourceId: { in: ids.toString() } }
+    })
+    const countPermission = permissionPaginate.meta.totalItem
+    if (countPermission > 0)
+      throw new ConflictException(
+        `Không thể xóa, các resource đã chọn đang được ${countPermission} permission sử dụng`
+      )
     await this.resourceRepository.delete({ id: In(ids) })
   }
 
@@ -68,7 +88,7 @@ export class ResourceService implements ICrudServices {
     return QueryHelper.buildReponse(result, totalItem, query)
   }
 
-  async findOne(id: number): Promise<any> {
+  async findOne(id: number) {
     const resource = await this.resourceRepository.findOne({ where: { id } })
     if (!resource) throw new NotFoundException('Không tìm thấy resource')
     return resource
